Add types and utility functions for tag management in Tags component
