+++ app/editor/src/features/content/form/components/tags/index.ts
@@ -1,2 +1,3 @@
 export * from './Tags';
 export * from './TagsContext';
+export * from './types';
